Refactor/be#36: api-server의 매칭 및 AI 로직을 각 모듈로 이동
