+++ backend/module-ai/src/main/java/com/team6/module/ai/support/RequestGuideCandidateProvider.java
@@ -1,4 +1,4 @@
-package com.team6.apiserver.service;
+package com.team6.module.ai.support;
 
 import com.team6.module.ai.dto.request.GuideRecommendRequest;
 import org.springframework.stereotype.Component;
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public class RequestGuideCandidateProvider implements GuideCandidateProvider {
+public class RequestGuideCandidateProvider implements com.team6.module.ai.support.GuideCandidateProvider {
 
     @Override
     public List<GuideRecommendRequest.GuideCandidateDto> getCandidates(
